We have changed the switch from Rental to Movie becuse it is a better idea to use a switch condition own's class data. Also we have relinked getCharge from Rental to the one into Movie which needs to be passed an int which indicated rental's duration. pag 35:
The first part of this problem is
that switch statement. It is a bad idea to do a switch based on an
attribute of another object. If you must use a switch statement, it should be on your own data, not
on someone else's

For this to work I had to pass in the length of the rental, which of course is data from
 the rental.
The method effectively uses two pieces of data, the length of the rental and the type of the movie.
Why do I prefer to pass the length of rental to the movie rather than the movie type to the rental?
It's because the proposed changes are all a
bout adding new types. Type information generally
tends to be more volatile. If I change the movie type, I want the least ripple effect, so I prefer to
calculate the charge within the movie.
I compiled the method into movie and then changed the
getCharge
 o
n rental to use the new
method